[bos #47355] integration of PyBatch

+++ products/compil_scripts/pybatch.sh
@@ -1,8 +1,10 @@
 #!/bin/bash
 
 echo "##########################################################################"
-echo "pytest $VERSION"
+echo "pybatch $VERSION"
 echo "##########################################################################"
+
+LINUX_DISTRIBUTION="$DIST_NAME$DIST_VERSION"
 
 fix_lib_path(){
 	mkdir -p $PRODUCT_INSTALL/lib/python$PYTHON_VERSION
@@ -31,23 +33,45 @@ fix_lib_path(){
 	fi
 }
 
-rm -rf "${BUILD_DIR}"
-mkdir "${BUILD_DIR}"
-cd "${BUILD_DIR}" || { echo "cd ${BUILD_DIR} fails"; exit 1; }
-cp -r "${SOURCE_DIR}"/* .
+if [ -d "${BUILD_DIR}" ]; then
+    rm -rf "${BUILD_DIR}"
+fi
+mkdir $BUILD_DIR
+mkdir -p ${BUILD_DIR}/cache/pip
+cd "${BUILD_DIR}"
 
 export PIP_DISABLE_PIP_VERSION_CHECK=1
 export PIP_CACHE_DIR=${BUILD_DIR}/cache/pip
 
+PIP_INSTALL_OPTIONS=
+PIP_INSTALL_OPTIONS+=" --no-binary :all:"
+if [ $LINUX_DISTRIBUTION != "DB11" ]; then
+    PIP_INSTALL_OPTIONS+=" --no-build-isolation"
+fi
+PIP_INSTALL_OPTIONS+=" --prefix=${PRODUCT_INSTALL}"
+PIP_INSTALL_OPTIONS+=" -vvv"
+
+
 echo
-echo "*** install with $PYTHONBIN -m pip install . --no-binary :all: --prefix=$PRODUCT_INSTALL --no-build-isolation -vvv"
-${PYTHONBIN} -m pip install . --no-binary :all: "pluggy" --prefix=$PRODUCT_INSTALL -vvv
+echo "INFO: running installation command"
+echo "*** ${PYTHONBIN} -m pip install ${SOURCE_DIR}[paramiko] ${PIP_INSTALL_OPTIONS}"
+${PYTHONBIN} -m pip install "${SOURCE_DIR}[paramiko]" ${PIP_INSTALL_OPTIONS}
 if [ $? -ne 0 ]; then
-echo "pip install pytest fails"
-exit 1
+    echo "pip install pybatch fails"
+    exit 4
 fi
 
 fix_lib_path
+
+#tests
+export PATH=${PRODUCT_INSTALL}/bin:${PATH}
+export PYTHONPATH=${PRODUCT_INSTALL}/lib/python${PYTHON_VERSION}/site-packages:${PYTHONPATH}
+
+${PYTHONBIN} -m pytest ${SOURCE_DIR}
+if [ $? -ne 0 ]; then
+    echo "ERROR: pybatch tests fails"
+    exit 5
+fi
 
 echo
 echo "########## END"
